# docs/swagger.yaml
basePath: /
definitions:
...
  dto.AttachmentListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.AttachmentResponse'
        type: array
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.AttachmentResponse:
    properties:
      category:
        example: RADIOLOGY
        type: string
      file_name:
        example: xray_chest.dcm
        type: string
      file_size:
        example: 4587210
        type: integer
      file_type:
        example: application/dicom
        type: string
      file_url:
        example: https://res.cloudinary.com/...
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      message:
        type: string
      metadata:
        additionalProperties: true
        type: object
      referral_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      rejected_at:
        example: "2026-04-11T19:55:00Z"
        type: string
      rejection_message:
        example: Metadata extraction failed
        type: string
      success:
        example: true
        type: boolean
      uploaded_at:
        type: string
      verification_status:
        example: VERIFIED
        type: string
    type: object
  dto.BaseResponse:
    properties:
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.BulkAttachmentRequest:
    properties:
      attachments:
        items:
          $ref: '#/definitions/dto.CreateAttachmentRequest'
        type: array
    required:
    - attachments
    type: object
...
  dto.CreateAttachmentRequest:
    properties:
      category:
        example: RADIOLOGY
        type: string
      file_name:
        example: xray_chest.dcm
        type: string
      file_size:
        example: 4587210
        type: integer
      file_type:
        example: application/dicom
        type: string
      file_url:
        example: https://res.cloudinary.com/...
        type: string
      public_id:
        example: referrals/hosp_123/ref_456/abc123
        type: string
    required:
    - file_name
    - file_type
    - file_url
    - public_id
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  dto.UploadSignatureResponse:
    properties:
      api_key:
        example: "123456789"
        type: string
      cloud_name:
        example: hospital-system
        type: string
      folder:
        example: temp/550e8400-e29b-41d4-a716-446655440000
        type: string
      message:
        type: string
      referral_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      signature:
        example: a92425642...
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: 1649684700
        type: integer
    type: object
...
info:
  contact: {}
  description: |-
    # Hospital Referral Hub API
    A national‑scale hospital referral management platform that digitises the entire patient‑transfer workflow, from initial doctor referral to final clinical outcome. All actions are governed by strict role‑based access controls and clinical governance rules.

    ---
    ## Referral Lifecycle
    DRAFT → SUBMITTED → (Liaison) UNDER_LIAISON_REVIEW → FORWARDED
    ↘ REJECTED_BY_LIAISON
    FORWARDED → (Specialist) UNDER_SPECIALIST_REVIEW → ACCEPTED / REJECTED_BY_SPECIALIST
    ACCEPTED → SCHEDULED → ASSIGNED → COMPLETED
    ↘ MISSED / RESCHEDULED / DECEASED

    ---
    ## Visibility Rules

    | Status | Visible To |
    |----------------------|-----------|
    | DRAFT / NEED_REVISION | Referring Doctor only |
    | SUBMITTED … FORWARDED | Liaison of the sender hospital |
    | FORWARDED … COMPLETED | Specialists of the target hospital |
    | ACCEPTED … RESCHEDULED| Receptionists of the target hospital |
    | All statuses | System Admins (global); MoH Analysts (aggregated dashboards, no raw clinical data) |

    ---
    ## Critical Business Rules

    - **ML Triage Gate**: A referral cannot be accepted without a severity score (set manually via `POST /specialist/referrals/{id}/triage-severity`).
...
    - **Walk‑in Restriction**: Walk‑ins can only be registered for referrals in ACCEPTED or SCHEDULED status.
...

    For detailed per‑endpoint rules, see the individual endpoint descriptions below.
...
paths:
...
  /api/v1/attachments/{id}:
    get:
      description: |-
        Retrieve metadata for a specific attachment by its ID.
        **Roles:** All authenticated roles with referral access.
        **Common Errors:**
        - 404 Not Found
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AttachmentResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Attachment Details
      tags:
      - Attachments
  /api/v1/attachments/{id}/verify:
    post:
      description: |-
        Immediately triggers metadata extraction and folder promotion for a specific attachment.
        **Roles:** SYSTEM_SUPER_ADMIN, HOSPITAL_ADMIN, DEPT_HEAD
        **State Transition:** Verified files are moved from `temp/` to permanent hospital storage. Failed files mark referral as NEED_REVISION.
        **Common Errors:**
        - 500 Internal Server Error
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AttachmentListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Manual Verification & Promotion
      tags:
      - Attachments
  /api/v1/attachments/signature:
    get:
      description: |-
        The first step in creating/updating a referral with attachments.
        1. Generates a unique **Pre-Minted Referral ID**.
        2. Provides a cryptographic signature for Cloudinary.
        3. Frontend MUST upload files to the folder path: `temp/{referral_id}/`.
        4. Use the returned `referral_id` when calling the Referral Creation API.
        **Roles:** REFERRING_DOCTOR
        **Common Errors:**
        - 500 Internal Server Error
      parameters:
      - description: Existing Referral ID (for updates)
        in: query
        name: referral_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UploadSignatureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Secure Upload Signature (Pre-Minted Flow)
      tags:
      - Attachments
...
  /api/v1/referrals/{id}/attachments:
    get:
      description: |-
        Get all attachments associated with a specific referral.
        **Roles:** All authenticated roles with referral access.
        **Common Errors:**
        - 404 Referral not found
      parameters:
      - description: Referral ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AttachmentListResponse'
      security:
      - BearerAuth: []
      summary: List Referral Attachments
      tags:
      - Attachments
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: |-
        Supports two modes for associating clinical data with a referral:
        1. **Hybrid/Bulk JSON**: Register multiple files already uploaded to Cloudinary.
        - Requires the `referral_id` pre-minted from the `/attachments/signature` endpoint.
        - Files MUST be at the `temp/{referral_id}/` path in Cloudinary.
        2. **Direct File**: Upload a single file (Max 20MB) directly to the backend.
        **Roles:** REFERRING_DOCTOR
        **Prerequisites:** Allowed only when referral status is DRAFT or NEED_REVISION.
        **Common Errors:**
        - 400 Invalid ID or file too large
        - 403 Forbidden (not the creator or invalid status)
      parameters:
      - description: Referral ID (Pre-minted or Existing)
        in: path
        name: id
        required: true
        type: string
      - description: Bulk JSON payload for Hybrid Flow
        in: body
        name: body
        schema:
          $ref: '#/definitions/dto.BulkAttachmentRequest'
      - description: Direct file upload payload
        in: formData
        name: file
        type: file
      - description: Category (e.g. RADIOLOGY, LAB_REPORT, DICOM_XRAY)
        in: formData
        name: category
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AttachmentListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register or Upload Referral Attachments
      tags:
      - Attachments
  /api/v1/referrals/{id}/attachments/{attachment_id}:
    delete:
      description: |-
        Remove attachment from referral.
        **Roles:** REFERRING_DOCTOR
        **Prerequisites:** Only allowed for Draft/NeedRevision status by the referring doctor.
        **Common Errors:**
        - 403 Forbidden (invalid status or not owner)
      parameters:
      - description: Referral ID
        in: path
        name: id
        required: true
        type: string
      - description: Attachment ID
        in: path
        name: attachment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Attachment From Referral
      tags:
      - Attachments
...
  /api/v1/referrals/{id}/verify-attachments:
    post:
      description: |-
        Triggers verification and Cloudinary promotion for ALL pending attachments of a specific referral.
        **Roles:** SYSTEM_SUPER_ADMIN, HOSPITAL_ADMIN, DEPT_HEAD, REFERRING_DOCTOR
        **State Transition:** Ensures the system moves the referral out of PENDING states before Liaison review.
        **Common Errors:**
        - 500 Internal Server Error
      parameters:
      - description: Referral ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Verify All Referral Attachments
      tags:
      - Attachments
...
securityDefinitions:
  BearerAuth:
    description: Enter your bearer token in the format **Bearer &lt;token&gt;**
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"